modify the place of the print so the header and the doc

+++ sort.h
@@ -29,6 +29,6 @@ void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
 void _swap(int *a, int *b);
 int lomutopartition(int array[], int low, int high);
-void quicksortrange(int array[], size_t low, size_t high, size_t size);
+void quicksortrange(int array[], int low, int high, size_t size);
 
 #endif /** _SORT_H_ */
